PP-686: gas estimation task (#245)
* feat: ported gas estimation tests to script
