+++ Readme.md
@@ -6,7 +6,7 @@ A secure transaction relay system to enable users to pay fees using ERC-20 token
 
 ## Description
 
-RIF Enveloping takes its inspiration from the [Gas Station Network (GSN) project](https://github.com/opengsn/gsn). GSN is a decentralized system that improves dApp usability without sacrificing security. In a nutshell, GSN abstracts away gas (used to pay transaction fees) to minimize onboarding and UX friction for dApps. With GSN, “gasless clients” can interact with smart contracts paying for gas with tokens instead of native-currency.
+RIF Relay takes its inspiration from the [Gas Station Network (GSN) project](https://github.com/opengsn/gsn). GSN is a decentralized system that improves dApp usability without sacrificing security. In a nutshell, GSN abstracts away gas (used to pay transaction fees) to minimize onboarding and UX friction for dApps. With GSN, “gasless clients” can interact with smart contracts paying for gas with tokens instead of native-currency.
 
 Our main objective is to provide the RSK ecosystem with the means to enable blockchain applications and end-users (wallet-apps) to pay for transaction fees using tokens (e.g. RIF tokens), and thereby remove the need to acquire RBTC in advance.
 
@@ -14,12 +14,11 @@ It is important to recall that - as a security measure - the version 1 contracts
 
 ## Modules
 
-Rif Relay is built in modules, the entire system is made up by 4 modules.
+Rif Relay is built in modules, the entire system is made up by 3 modules.
 
-1. [Rif Relay Contracts](https://github.com/anarancio/rif-relay-contracts) contains all the contracts used by the Rif Relay System.
-2. [Rif Relay Common](https://github.com/infuy/rif-relay-common) it’s a library that contains all the common code used by other modules.
-3. [Rif Relay Client](https://github.com/infuy/rif-relay-client) contains a library to interact with the relay server.
-4. [Rif Relay Server](https://github.com/infuy/rif-relay-server) has all the relay server code, you can use it as a library or run the server directly from there.
+1. [Rif Relay Contracts](https://github.com/rsksmart/rif-relay-contracts) contains all the contracts used by the Rif Relay System.
+2. [Rif Relay Client](https://github.com/rsksmart/rif-relay-client) contains a library to interact with the relay server.
+3. [Rif Relay Server](https://github.com/rsksmart/rif-relay-server) has all the relay server code. You can run the server directly from there.
 
 Each module has instructions for development and usage.
 
@@ -36,5 +35,15 @@ that contains the test code inside will execute only that test and make the veri
 **Important Note**: when you run the tests you will see folders like `contracts`, `migrations` and some other files
 appearing during the tests and disappearing after, that’s because we need to have the contracts
 on the same context as this repository during tests, this is a normal behaviour. 
+
+## Scripts
+
+This repository contains a script that analyzes gas consumption for relay requests. It prints different metrics related to gas usage when relaying a transaction.
+To run it, execute the command:
+
+`npm run analyze:gas <NETWORK_NAME>`
+
+where:
+- `<NETWORK_NAME>` is a required parameter for the network name, taken from the `hardhat.config.ts`, and must correspond to a valid RSK Node (commonly, regtest is used). Do not use this with the default network (hardhat) as it will fail.
 
 [Deprecated Doc](docs/README.md)
