+++ package.json
@@ -72,7 +72,8 @@
     "lint:fix": "npm run lint -- --fix",
     "prepare": "utils/prepare.npm",
     "test": "ALLOW_CONFIG_MUTATIONS=true NODE_CONFIG_DIR='node_modules/@rsksmart/rif-relay-server/config:config' TEST=true REPORT_GAS=true hardhat test --typecheck",
-    "test:docker": "utils/test"
+    "test:docker": "utils/test",
+    "analyze:gas": "npx hardhat run scripts/RelayHubGasEstimation.ts --network"
   },
   "lint-staged": {
     "*.ts": "eslint --cache --fix",
